Cleaning up typos and grammar

+++ SHE_PPT/python/SHE_PPT/table_formats/she_common_calibration.py
@@ -24,7 +24,6 @@ __updated__ = "2021-08-12"
 from ..logging import getLogger
 from ..table_utility import SheTableFormat, SheTableMeta
 
-
 fits_version = "8.0"
 fits_def = "she.commonCalibration"
 
@@ -52,101 +51,101 @@ class SheCommonCalibrationFormat(SheTableFormat):
         # Table column labels and properties
 
         self.f0_min = self.set_column_properties(
-            "SHE_CC_F0_MIN", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F0_MIN", dtype = ">f4", fits_dtype = "E")
         self.f0_max = self.set_column_properties(
-            "SHE_CC_F0_MAX", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F0_MAX", dtype = ">f4", fits_dtype = "E")
         self.f1_min = self.set_column_properties(
-            "SHE_CC_F1_MIN", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F1_MIN", dtype = ">f4", fits_dtype = "E")
         self.f1_max = self.set_column_properties(
-            "SHE_CC_F1_MAX", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F1_MAX", dtype = ">f4", fits_dtype = "E")
         self.f2_min = self.set_column_properties(
-            "SHE_CC_F2_MIN", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F2_MIN", dtype = ">f4", fits_dtype = "E")
         self.f2_max = self.set_column_properties(
-            "SHE_CC_F2_MAX", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F2_MAX", dtype = ">f4", fits_dtype = "E")
         self.f3_min = self.set_column_properties(
-            "SHE_CC_F3_MIN", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F3_MIN", dtype = ">f4", fits_dtype = "E")
         self.f3_max = self.set_column_properties(
-            "SHE_CC_F3_MAX", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F3_MAX", dtype = ">f4", fits_dtype = "E")
         self.f4_min = self.set_column_properties(
-            "SHE_CC_F4_MIN", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F4_MIN", dtype = ">f4", fits_dtype = "E")
         self.f4_max = self.set_column_properties(
-            "SHE_CC_F4_MAX", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F4_MAX", dtype = ">f4", fits_dtype = "E")
         self.f5_min = self.set_column_properties(
-            "SHE_CC_F5_MIN", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F5_MIN", dtype = ">f4", fits_dtype = "E")
         self.f5_max = self.set_column_properties(
-            "SHE_CC_F5_MAX", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F5_MAX", dtype = ">f4", fits_dtype = "E")
         self.f6_min = self.set_column_properties(
-            "SHE_CC_F6_MIN", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F6_MIN", dtype = ">f4", fits_dtype = "E")
         self.f6_max = self.set_column_properties(
-            "SHE_CC_F6_MAX", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F6_MAX", dtype = ">f4", fits_dtype = "E")
         self.f7_min = self.set_column_properties(
-            "SHE_CC_F7_MIN", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F7_MIN", dtype = ">f4", fits_dtype = "E")
         self.f7_max = self.set_column_properties(
-            "SHE_CC_F7_MAX", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F7_MAX", dtype = ">f4", fits_dtype = "E")
         self.f8_min = self.set_column_properties(
-            "SHE_CC_F8_MIN", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F8_MIN", dtype = ">f4", fits_dtype = "E")
         self.f8_max = self.set_column_properties(
-            "SHE_CC_F8_MAX", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F8_MAX", dtype = ">f4", fits_dtype = "E")
         self.f9_min = self.set_column_properties(
-            "SHE_CC_F9_MIN", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F9_MIN", dtype = ">f4", fits_dtype = "E")
         self.f9_max = self.set_column_properties(
-            "SHE_CC_F9_MAX", dtype=">f4", fits_dtype="E")
+            "SHE_CC_F9_MAX", dtype = ">f4", fits_dtype = "E")
         self.i0_min = self.set_column_properties(
-            "SHE_CC_I0_MIN", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I0_MIN", dtype = ">i8", fits_dtype = "K")
         self.i0_max = self.set_column_properties(
-            "SHE_CC_I0_MAX", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I0_MAX", dtype = ">i8", fits_dtype = "K")
         self.i1_min = self.set_column_properties(
-            "SHE_CC_I1_MIN", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I1_MIN", dtype = ">i8", fits_dtype = "K")
         self.i1_max = self.set_column_properties(
-            "SHE_CC_I1_MAX", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I1_MAX", dtype = ">i8", fits_dtype = "K")
         self.i2_min = self.set_column_properties(
-            "SHE_CC_I2_MIN", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I2_MIN", dtype = ">i8", fits_dtype = "K")
         self.i2_max = self.set_column_properties(
-            "SHE_CC_I2_MAX", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I2_MAX", dtype = ">i8", fits_dtype = "K")
         self.i3_min = self.set_column_properties(
-            "SHE_CC_I3_MIN", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I3_MIN", dtype = ">i8", fits_dtype = "K")
         self.i3_max = self.set_column_properties(
-            "SHE_CC_I3_MAX", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I3_MAX", dtype = ">i8", fits_dtype = "K")
         self.i4_min = self.set_column_properties(
-            "SHE_CC_I4_MIN", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I4_MIN", dtype = ">i8", fits_dtype = "K")
         self.i4_max = self.set_column_properties(
-            "SHE_CC_I4_MAX", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I4_MAX", dtype = ">i8", fits_dtype = "K")
         self.i5_min = self.set_column_properties(
-            "SHE_CC_I5_MIN", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I5_MIN", dtype = ">i8", fits_dtype = "K")
         self.i5_max = self.set_column_properties(
-            "SHE_CC_I5_MAX", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I5_MAX", dtype = ">i8", fits_dtype = "K")
         self.i6_min = self.set_column_properties(
-            "SHE_CC_I6_MIN", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I6_MIN", dtype = ">i8", fits_dtype = "K")
         self.i6_max = self.set_column_properties(
-            "SHE_CC_I6_MAX", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I6_MAX", dtype = ">i8", fits_dtype = "K")
         self.i7_min = self.set_column_properties(
-            "SHE_CC_I7_MIN", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I7_MIN", dtype = ">i8", fits_dtype = "K")
         self.i7_max = self.set_column_properties(
-            "SHE_CC_I7_MAX", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I7_MAX", dtype = ">i8", fits_dtype = "K")
         self.i8_min = self.set_column_properties(
-            "SHE_CC_I8_MIN", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I8_MIN", dtype = ">i8", fits_dtype = "K")
         self.i8_max = self.set_column_properties(
-            "SHE_CC_I8_MAX", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I8_MAX", dtype = ">i8", fits_dtype = "K")
         self.i9_min = self.set_column_properties(
-            "SHE_CC_I9_MIN", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I9_MIN", dtype = ">i8", fits_dtype = "K")
         self.i9_max = self.set_column_properties(
-            "SHE_CC_I9_MAX", dtype=">i8", fits_dtype="K")
+            "SHE_CC_I9_MAX", dtype = ">i8", fits_dtype = "K")
         self.m1 = self.set_column_properties(
-            "SHE_CC_M1", dtype=">f4", fits_dtype="E")
+            "SHE_CC_M1", dtype = ">f4", fits_dtype = "E")
         self.m1_err = self.set_column_properties(
-            "SHE_CC_M1_ERR", dtype=">f4", fits_dtype="E")
+            "SHE_CC_M1_ERR", dtype = ">f4", fits_dtype = "E")
         self.m2 = self.set_column_properties(
-            "SHE_CC_M2", dtype=">f4", fits_dtype="E")
+            "SHE_CC_M2", dtype = ">f4", fits_dtype = "E")
         self.m2_err = self.set_column_properties(
-            "SHE_CC_M2_ERR", dtype=">f4", fits_dtype="E")
+            "SHE_CC_M2_ERR", dtype = ">f4", fits_dtype = "E")
         self.c1 = self.set_column_properties(
-            "SHE_CC_C1", dtype=">f4", fits_dtype="E")
+            "SHE_CC_C1", dtype = ">f4", fits_dtype = "E")
         self.c1_err = self.set_column_properties(
-            "SHE_CC_C1_ERR", dtype=">f4", fits_dtype="E")
+            "SHE_CC_C1_ERR", dtype = ">f4", fits_dtype = "E")
         self.c2 = self.set_column_properties(
-            "SHE_CC_C2", dtype=">f4", fits_dtype="E")
+            "SHE_CC_C2", dtype = ">f4", fits_dtype = "E")
         self.c2_err = self.set_column_properties(
-            "SHE_CC_C2_ERR", dtype=">f4", fits_dtype="E")
+            "SHE_CC_C2_ERR", dtype = ">f4", fits_dtype = "E")
 
         self._finalize_init()
 
@@ -154,5 +153,5 @@ class SheCommonCalibrationFormat(SheTableFormat):
 # Define an instance of this object that can be imported
 common_calibration_table_format = SheCommonCalibrationFormat()
 
-# And a convient alias for it
+# And a convenient alias for it
 tf = common_calibration_table_format
