+++ SHE_PPT/python/SHE_PPT/table_formats/she_lensmc_measurements.py
@@ -29,7 +29,6 @@ from ..table_formats.mer_final_catalog import tf as mfc_tf
 from ..table_formats.she_measurements import SheMeasurementsMeta, SheMeasurementsFormat
 from ..table_utility import is_in_format, init_table
 
-
 fits_version = "8.0"
 fits_def = "she.lensmcMeasurements"
 
@@ -52,23 +51,23 @@ def set_lensmc_column_properties(tf):
         reuse with the she_lensmc_tu_matched table.
     """
     tf.snr_err = tf.set_column_properties(
-        "SHE_LENSMC_SNR_ERR", dtype=">f4", fits_dtype="E")
+        "SHE_LENSMC_SNR_ERR", dtype = ">f4", fits_dtype = "E")
     tf.bulge_frac = tf.set_column_properties(
-        "SHE_LENSMC_BULGE_FRAC", dtype=">f4", fits_dtype="E")
+        "SHE_LENSMC_BULGE_FRAC", dtype = ">f4", fits_dtype = "E")
     tf.bulge_frac_err = tf.set_column_properties(
-        "SHE_LENSMC_BULGE_FRAC_ERR", dtype=">f4", fits_dtype="E")
+        "SHE_LENSMC_BULGE_FRAC_ERR", dtype = ">f4", fits_dtype = "E")
     tf.gal_pvalue = tf.set_column_properties(
-        "SHE_LENSMC_GAL_PVALUE", dtype=">f4", fits_dtype="E")
+        "SHE_LENSMC_GAL_PVALUE", dtype = ">f4", fits_dtype = "E")
     tf.chi2 = tf.set_column_properties(
-        "SHE_LENSMC_CHI2", dtype=">f4", fits_dtype="E")
+        "SHE_LENSMC_CHI2", dtype = ">f4", fits_dtype = "E")
     tf.dof = tf.set_column_properties(
-        "SHE_LENSMC_DOF", dtype=">i4", fits_dtype="K")
+        "SHE_LENSMC_DOF", dtype = ">i4", fits_dtype = "K")
     tf.acc = tf.set_column_properties(
-        "SHE_LENSMC_ACCEPTANCE", dtype=">f4", fits_dtype="E")
+        "SHE_LENSMC_ACCEPTANCE", dtype = ">f4", fits_dtype = "E")
     tf.m1_ical = tf.set_column_properties(
-        "SHE_LENSMC_M1_ICAL", dtype=">f4", fits_dtype="E")
+        "SHE_LENSMC_M1_ICAL", dtype = ">f4", fits_dtype = "E")
     tf.m2_ical = tf.set_column_properties(
-        "SHE_LENSMC_M2_ICAL", dtype=">f4", fits_dtype="E")
+        "SHE_LENSMC_M2_ICAL", dtype = ">f4", fits_dtype = "E")
 
 
 class SheLensMcMeasurementsFormat(SheMeasurementsFormat):
@@ -78,9 +77,8 @@ class SheLensMcMeasurementsFormat(SheMeasurementsFormat):
     """
 
     def __init__(self):
-
         # Inherit format from parent class, and save it in separate dicts so we can properly adjust column names
-        super().__init__(SheLensMcMeasurementsMeta(), finalize=False)
+        super().__init__(SheLensMcMeasurementsMeta(), finalize = False)
 
         self.setup_child_table_format(child_label)
 
@@ -100,18 +98,18 @@ class SheLensMcMeasurementsFormat(SheMeasurementsFormat):
 # Define an instance of this object that can be imported
 lensmc_measurements_table_format = SheLensMcMeasurementsFormat()
 
-# And a convient alias for it
+# And a convenient alias for it
 tf = lensmc_measurements_table_format
 
 
 def make_lensmc_measurements_table_header(
-        model_hash=None,
-        model_seed=None,
-        noise_seed=None,
-        observation_id=None,
-        pointing_id=None,
-        observation_time=None,
-        tile_id=None,):
+        model_hash = None,
+        model_seed = None,
+        noise_seed = None,
+        observation_id = None,
+        pointing_id = None,
+        observation_time = None,
+        tile_id = None, ):
     """
         @brief Generate a header for a shear estimates table.
 
@@ -145,17 +143,17 @@ def make_lensmc_measurements_table_header(
     return header
 
 
-def initialise_lensmc_measurements_table(mer_final_catalog=None,
-                                         size=None,
-                                         optional_columns=None,
-                                         init_cols=None,
-                                         model_hash=None,
-                                         model_seed=None,
-                                         noise_seed=None,
-                                         observation_id=None,
-                                         pointing_id=None,
-                                         observation_time=None,
-                                         tile_id=None,
+def initialise_lensmc_measurements_table(mer_final_catalog = None,
+                                         size = None,
+                                         optional_columns = None,
+                                         init_cols = None,
+                                         model_hash = None,
+                                         model_seed = None,
+                                         noise_seed = None,
+                                         observation_id = None,
+                                         pointing_id = None,
+                                         observation_time = None,
+                                         tile_id = None,
                                          ):
     """
         @brief Initialise a shear estimates table based on a detections table, with the
@@ -172,7 +170,7 @@ def initialise_lensmc_measurements_table(mer_final_catalog=None,
     """
 
     assert (mer_final_catalog is None) or (
-        is_in_format(mer_final_catalog, mfc_tf, strict=False))
+        is_in_format(mer_final_catalog, mfc_tf, strict = False))
 
     if optional_columns is None:
         optional_columns = []
@@ -182,16 +180,16 @@ def initialise_lensmc_measurements_table(mer_final_catalog=None,
             if colname not in tf.all:
                 raise ValueError("Invalid optional column name: " + colname)
 
-    lensmc_measurements_table = init_table(tf, optional_columns=optional_columns, init_cols=init_cols, size=size)
+    lensmc_measurements_table = init_table(tf, optional_columns = optional_columns, init_cols = init_cols, size = size)
 
     lensmc_measurements_table.meta = make_lensmc_measurements_table_header(
-        model_hash=model_hash,
-        model_seed=model_seed,
-        noise_seed=noise_seed,
-        observation_id=observation_id,
-        pointing_id=pointing_id,
-        observation_time=observation_time,
-        tile_id=tile_id)
+        model_hash = model_hash,
+        model_seed = model_seed,
+        noise_seed = noise_seed,
+        observation_id = observation_id,
+        pointing_id = pointing_id,
+        observation_time = observation_time,
+        tile_id = tile_id)
 
     assert is_in_format(lensmc_measurements_table, tf)
