+++ SHE_PPT/python/SHE_PPT/table_formats/she_psf_model_image.py
@@ -27,7 +27,6 @@ from ..constants.fits import FITS_VERSION_LABEL, FITS_DEF_LABEL, EXTNAME_LABEL, 
 from ..logging import getLogger
 from ..table_utility import SheTableFormat, SheTableMeta
 
-
 fits_version = "8.0"
 fits_def = "she.psfModelImage.shePsfC"
 
@@ -54,18 +53,17 @@ class ShePsfModelImageMeta(SheTableMeta):
     field_time: str = "FLD_TIME"
 
     def __init__(self):
-
-        super().__init__(comments=OrderedDict(((self.fits_version, None),
-                                               (self.fits_def, None),
-                                               (self.extname, PSF_CAT_TAG),
-                                               (self.calibration_product, None),
-                                               (self.calibration_time, None),
-                                               (self.field_product, None),
-                                               (self.field_time, None)
-                                               )))
+        super().__init__(comments = OrderedDict(((self.fits_version, None),
+                                                 (self.fits_def, None),
+                                                 (self.extname, PSF_CAT_TAG),
+                                                 (self.calibration_product, None),
+                                                 (self.calibration_time, None),
+                                                 (self.field_product, None),
+                                                 (self.field_time, None)
+                                                 )))
 
     def init_meta(self, **kwargs: str):
-        return super().init_meta(extname=PSF_CAT_TAG,
+        return super().init_meta(extname = PSF_CAT_TAG,
                                  **kwargs)
 
 
@@ -81,29 +79,29 @@ class ShePsfModelImageFormat(SheTableFormat):
         # Column names and info
 
         self.ID = self.set_column_properties(
-            "OBJECT_ID", dtype=">i8", fits_dtype="K")
+            "OBJECT_ID", dtype = ">i8", fits_dtype = "K")
 
         self.template = self.set_column_properties(
-            "SHE_PSF_SED_TEMPLATE", dtype=">i8", fits_dtype="K")
+            "SHE_PSF_SED_TEMPLATE", dtype = ">i8", fits_dtype = "K")
         self.bulge_index = self.set_column_properties(
-            "SHE_PSF_BULGE_INDEX", dtype=">i4", fits_dtype="J")
+            "SHE_PSF_BULGE_INDEX", dtype = ">i4", fits_dtype = "J")
         self.disk_index = self.set_column_properties(
-            "SHE_PSF_DISK_INDEX", dtype=">i4", fits_dtype="J")
+            "SHE_PSF_DISK_INDEX", dtype = ">i4", fits_dtype = "J")
         self.image_x = self.set_column_properties(
-            "SHE_PSF_IMAGE_X", dtype=">i2", fits_dtype="I")
+            "SHE_PSF_IMAGE_X", dtype = ">i2", fits_dtype = "I")
         self.image_y = self.set_column_properties(
-            "SHE_PSF_IMAGE_Y", dtype=">i2", fits_dtype="I")
+            "SHE_PSF_IMAGE_Y", dtype = ">i2", fits_dtype = "I")
         self.x = self.set_column_properties(
-            "SHE_PSF_X", dtype=">f4", fits_dtype="E")
+            "SHE_PSF_X", dtype = ">f4", fits_dtype = "E")
         self.y = self.set_column_properties(
-            "SHE_PSF_Y", dtype=">f4", fits_dtype="E")
+            "SHE_PSF_Y", dtype = ">f4", fits_dtype = "E")
 
         self.calibration_time = self.set_column_properties(
-            "SHE_PSF_CALIB_TIME", dtype="str", fits_dtype="A", length=20)
+            "SHE_PSF_CALIB_TIME", dtype = "str", fits_dtype = "A", length = 20)
         self.field_time = self.set_column_properties(
-            "SHE_PSF_FIELD_TIME", dtype="str", fits_dtype="A", length=20)
+            "SHE_PSF_FIELD_TIME", dtype = "str", fits_dtype = "A", length = 20)
         self.qual_flag = self.set_column_properties(
-            "SHE_PSF_QUAL_FLAG", dtype=">i4", fits_dtype="J")
+            "SHE_PSF_QUAL_FLAG", dtype = ">i4", fits_dtype = "J")
 
         self._finalize_init()
 
@@ -111,5 +109,5 @@ class ShePsfModelImageFormat(SheTableFormat):
 # Define an instance of this object that can be imported
 psf_table_format = ShePsfModelImageFormat()
 
-# And a convient alias for it
+# And a convenient alias for it
 tf = psf_table_format
